refactor(model): Extend further abstracted
