+++ lib/utils/model/extend.js
@@ -77,18 +77,7 @@ module.exports = function extendModel() {
   this.models[modelName] = model;
 
   // Bind for nestedOperations
-  moduleLdr.on('routesInitialized', function() {
-    debug( 'Parsing templated event handlers...' );
-    Object.keys( Klass ).forEach( function( propName ) {
-      if ( propName.indexOf( ' ' ) !== -1 || utils.model.helpers.eventNames.indexOf( propName ) !== -1 ) {
-        var parts       = propName.split( ' ' )
-          , resource    = parts.length === 2 ? parts.shift() : modelName + 'Model'
-          , eventName   = parts.shift();
-
-        injector.getInstance( resource ).on( eventName, model.callback( propName ) );
-      }
-    });
-  });
+  utils.model.helpers.nestedOperations.apply(this, [Klass, modelName, model, debug]);
 
   return model;
 };
